requirements.txt
kafka-python-ng==2.2.3
pyspark==3.5.0
great-expectations==0.18.9
pandas==2.1.4
pyarrow==14.0.2
fastparquet==2023.10.1
prometheus-client==0.19.0
python-dotenv==1.0.0     
tenacity==8.2.3
loguru==0.7.2
pytest==7.4.3
pytest-mock==3.12.0